Fix data types.

+++ include/imu.h
@@ -8,10 +8,10 @@
 // Used to store the imu data
 struct imu_data
 {
-    double heading;
-    double gyro_z;
-    double accel_x;
-    double accel_y;
+    float heading;
+    float gyro_z;
+    float accel_x;
+    float accel_y;
 };
 
 // Starts imu and ensure fusion starts and no errors occur, returns true if start okay, false if not.
@@ -25,7 +25,7 @@ void displaySensorStatus();
 void imu_task();
 
 // Getter functions to expose imu data
-double imu_get_heading();
-double imu_get_gyro_z();
+float imu_get_heading();
+float imu_get_gyro_z();
 
 #endif // ROBOCUP_IMU_H
